Add posibility to use multiple Halls

+++ DotnetLab2.cs
@@ -9,38 +9,7 @@ public class Lab2 : UserCustom.Lab
 {
     public void Entry()
     {
-        Cinema cinema = new Cinema();
-        while(true)
-        {
-            Console.Write("Current movie: {0}\nSeats count: {1}\nFree seats count: {2}\n",
-                cinema.ticketsContext.Hall.CurrentMovie.MovieName, cinema.ticketsContext.Hall.SeatsCount, cinema.ticketsContext.Hall.FreeSeatsCount);
-            Console.Write("1: Add ticket \n2: Choose movie\n3: Add movie\nq: exit\n");
-            string? line = Console.ReadLine();
-            if (line == "q") break;
-            switch (line) {
-                case "1":
-                    Console.WriteLine("Choose movie:");
-                    foreach(var movie in cinema.ticketsContext.Movies)
-                    {
-                        Console.WriteLine(movie.MovieName);
-                    }
-                    cinema.AddTicket(Console.ReadLine());
-                    break;
-                case "2":
-                    Console.WriteLine("Choose movie:");
-                    foreach (var movie in cinema.ticketsContext.Movies)
-                    {
-                        Console.WriteLine(movie.MovieName);
-                    }
-                    cinema.UpdateHallMovieSession(Console.ReadLine());
-                    break;
-                case "3":
-                    Console.WriteLine("Print movie:");
-                    cinema.AddMovie(Console.ReadLine());
-                    break;
-                default: break;
-            }
-        }
+
     }
 }
 
@@ -71,8 +40,14 @@ public class Cinema
 
         ticketsContext.MovieTickets.RemoveRange(from ticket in ticketsContext.MovieTickets where ticket.MovieInfo == movie select ticket);
 
-        ticketsContext.Hall.CurrentMovie = Movie.EmptyMovie;
-
+        foreach (var hall in ticketsContext.Halls)
+        {
+            if (hall.CurrentMovie == movie)
+            {
+                hall.CurrentMovie = Movie.EmptyMovie;
+                hall.FreeSeatsCount = hall.SeatsCount;
+            }
+        }
 
         ticketsContext.Movies.Remove(movie);
         ticketsContext.SaveChanges();
@@ -86,32 +61,57 @@ public class Cinema
         ticketsContext.SaveChanges();
     }
 
+    public void AddTicket(string _MovieName, int HallNumber)
+    {
+        Movie? movie = (from movieInfo in ticketsContext.Movies where movieInfo.MovieName == _MovieName select movieInfo).FirstOrDefault();
+        if (movie is null) return;
+
+        Hall? choosenHall = (from hall in ticketsContext.Halls where hall.HallNumber == HallNumber select hall).FirstOrDefault();
+        if (choosenHall is null) return;
+
+        if (choosenHall.CurrentMovie != movie) return;
+
+        --choosenHall.FreeSeatsCount;
+
+        ticketsContext.MovieTickets.Add(
+            new MovieTicket
+            {
+                MovieInfo = movie,
+                HallInfo = choosenHall
+            }
+        );
+        ticketsContext.SaveChanges();
+    }
     public void AddTicket(string _MovieName)
     {
         Movie? movie = (from movieInfo in ticketsContext.Movies where movieInfo.MovieName == _MovieName select movieInfo).FirstOrDefault();
         if (movie is null) return;
 
-        if (ticketsContext.Hall.CurrentMovie != movie) return;
+        Hall? choosenHall = (from hall in ticketsContext.Halls where hall.CurrentMovie == movie select hall).FirstOrDefault();
+        if (choosenHall is null) return;
 
-
-        --ticketsContext.Hall.FreeSeatsCount;
+        --choosenHall.FreeSeatsCount;
 
         ticketsContext.MovieTickets.Add(
             new MovieTicket
             {
-                MovieInfo = movie
+                MovieInfo = movie,
+                HallInfo = choosenHall
             }
         );
         ticketsContext.SaveChanges();
     }
 
-    public void UpdateHallMovieSession(string _MovieName)
+    public void UpdateHallMovieSession(int HallNumber, string _MovieName)
     {
         Movie? movie = (from movieInfo in ticketsContext.Movies where movieInfo.MovieName == _MovieName select movieInfo).FirstOrDefault();
         if (movie is null) return;
 
-        ticketsContext.Hall.CurrentMovie = movie;
-        ticketsContext.Hall.FreeSeatsCount = ticketsContext.Hall.SeatsCount - (from ticket in ticketsContext.MovieTickets where ticket.MovieInfo == movie select ticket).Count();
+        Hall? choosenHall = (from hall in ticketsContext.Halls where hall.HallNumber == HallNumber select hall).FirstOrDefault();
+        if (choosenHall is null) return;
+
+        choosenHall.CurrentMovie = movie;
+        choosenHall.FreeSeatsCount = choosenHall.SeatsCount - (from ticket in ticketsContext.MovieTickets where ticket.MovieInfo == movie && ticket.HallInfo == choosenHall select ticket).Count();
         ticketsContext.MovieTickets.RemoveRange(from ticket in ticketsContext.MovieTickets where ticket.MovieInfo == movie select ticket);
         ticketsContext.SaveChanges();
     }
@@ -125,7 +125,7 @@ public class MovieTicketsContext : DbContext
     public DbSet<MovieTicket> MovieTickets { get; set; }
     public DbSet<Movie> Movies { get; set; }
     public DbSet<Hall> Halls { get; set; }
-    public Hall Hall { get => (from hall in Halls select hall).FirstOrDefault(); }
+    public Hall? Hall { get => (from hall in Halls select hall).FirstOrDefault(); }
 
     public string DbPath { get; }
 
@@ -146,6 +146,8 @@ public class MovieTicketsContext : DbContext
         modelBuilder.Entity<MovieTicket>().HasOne(ticket => ticket.MovieInfo).WithMany(movie => movie.Tickets).IsRequired();
         modelBuilder.Entity<Hall>().HasKey(hall => hall.HallNumber);
         modelBuilder.Entity<Hall>().HasOne(hall => hall.CurrentMovie).WithMany().IsRequired();
+        modelBuilder.Entity<MovieTicket>().HasOne(ticket => ticket.HallInfo).WithMany().IsRequired();
+        modelBuilder.Entity<MovieTicket>().HasKey(ticket => ticket.Id);
     }
 
     protected override void OnConfiguring(DbContextOptionsBuilder options)
@@ -160,6 +162,7 @@ public class MovieTicket
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public int Id { get; set; }
     public Movie MovieInfo { get; set; } = Movie.EmptyMovie;
+    public Hall HallInfo { get; set; } = Hall.EmptyHall;
 }
 
 public class Movie
@@ -173,11 +176,12 @@ public class Movie
 
 public class Hall
 {
+    public static Hall EmptyHall = new Hall();
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    public int HallNumber { get; set; }
+    public int HallNumber { get; set; } = 0;
     public Movie CurrentMovie = Movie.EmptyMovie;
-    public int SeatsCount { get; set; }
-    public int FreeSeatsCount { get; set; }
+    public int SeatsCount { get; set; } = 100;
+    public int FreeSeatsCount { get; set; } = 100;
 }
